@@ -1,6 +1,6 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
   <PropertyGroup>
-    <TargetFramework>netcoreapp2.0</TargetFramework>
+    <TargetFramework>netcoreapp2.1</TargetFramework>
     <Version>2.0.0</Version>
     <Authors>OptimaJet, LLC</Authors>
     <Company>OptimaJet, LLC</Company>
@@ -12,10 +12,10 @@
     <Copyright>OptimaJet DWKit 2018</Copyright>
   </PropertyGroup>
   <ItemGroup>
-    <PackageReference Include="DWKit-Core" Version="2.3.2" />
-    <PackageReference Include="DWKit-ProviderForMSSQL" Version="2.3.2" />
-    <PackageReference Include="DWKit-ProviderForPostgreSQL" Version="2.3.2" />
-    <PackageReference Include="DWKit-Security" Version="2.3.2" />
+    <PackageReference Include="DWKit-Core" Version="2.5.3" />
+    <PackageReference Include="DWKit-ProviderForMSSQL" Version="2.5.0" />
+    <PackageReference Include="DWKit-ProviderForPostgreSQL" Version="2.5.0" />
+    <PackageReference Include="DWKit-Security" Version="2.5.0" />
     <PackageReference Include="Microsoft.AspNetCore.Http.Abstractions" Version="2.1.1" />
     <PackageReference Include="Microsoft.CSharp" Version="4.4.1" />
     <PackageReference Include="Microsoft.Extensions.Configuration.Abstractions" Version="2.1.1" />
